Better multithreading support. Added rplObject.
Multithreading is much better.  Debugging now works when the Python
execution occurs in a different thread.  Also, the Python execution
process can be terminated on command (if multithreaded, of course).
Implemented the rplObject class, which allows the user to simply create
a class that derives from it and not have to worry about instantiation
in Python.  Python class registration is not automated yet; a feature
for this may be added in the future.  The rplObject class also has
built-in locking features, so all you have to do is create an instance
of rplScopedLock inside the block you wish to lock.  Unlocking occurs
automatically.